extract tracing logic to it's own plugin
