+++ projects/gloo/pkg/plugins/tracing/tracing_suite_test.go
@@ -1,0 +1,13 @@
+package tracing_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestTracing(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Tracing Suite")
+}
